Make thread support optional
This is a refactoring such that the core sfmt random number generator
only depends on rand_core, with no dependency on rand. The rand crate is
required for the thread support though, so to ensure that this
functionality is still available I have included it as an optional
feature.

Removing the requirement of rand did additionally require that some of
the tests were modified to use the smaller range of functionality of the
rand_core crate.